fix(test): 拆分消息流式测试并清理品牌残留
